requirements.txt
openai
langchain
faiss-cpu
streamlit
PyPDF2
# pdf2docx
sentence_transformers